edit chat type for handlers

+++ bot/handlers/commands/me.py
@@ -6,9 +6,11 @@ from bot.utils.messages import get_me_message
 from logger import logger
 
 
-@dp.message_handler(Command('me'), chat_type=types.ChatType.SUPERGROUP)
+@dp.message_handler(Command('me'), chat_type=[types.ChatType.SUPERGROUP, types.ChatType.GROUP])
 async def me_handler(message: types.Message):
     logger.info(f"{message.from_id} {message.text}")
+    if not message.get_args():
+        return
     await message.answer(get_me_message(message), parse_mode=types.ParseMode.MARKDOWN_V2)
     try:
         await message.delete()
